move client to pkg/client
