+++ actions/actions.go
@@ -18,8 +18,8 @@ package actions
 import (
 	"fmt"
 
-	"github.com/ksonnet/ksonnet/client"
 	"github.com/ksonnet/ksonnet/pkg/app"
+	"github.com/ksonnet/ksonnet/pkg/client"
 	"github.com/pkg/errors"
 	"github.com/spf13/afero"
 )
