+++ actions/delete_test.go
@@ -18,8 +18,8 @@ package actions
 import (
 	"testing"
 
-	"github.com/ksonnet/ksonnet/client"
 	amocks "github.com/ksonnet/ksonnet/pkg/app/mocks"
+	"github.com/ksonnet/ksonnet/pkg/client"
 	"github.com/ksonnet/ksonnet/pkg/cluster"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
