+++ actions/show.go
@@ -19,8 +19,8 @@ import (
 	"io"
 	"os"
 
-	"github.com/ksonnet/ksonnet/client"
 	"github.com/ksonnet/ksonnet/pkg/app"
+	"github.com/ksonnet/ksonnet/pkg/client"
 	"github.com/ksonnet/ksonnet/pkg/cluster"
 )
