+++ cmd/init.go
@@ -24,7 +24,7 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/ksonnet/ksonnet/actions"
-	"github.com/ksonnet/ksonnet/client"
+	"github.com/ksonnet/ksonnet/pkg/client"
 	"github.com/spf13/cobra"
 )
