+++ cmd/validate.go
@@ -21,7 +21,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/ksonnet/ksonnet/actions"
-	"github.com/ksonnet/ksonnet/client"
+	"github.com/ksonnet/ksonnet/pkg/client"
 )
 
 const (
