+++ pkg/client/testdata/swagger.json
