+++ pkg/cluster/cluster.go
@@ -18,8 +18,8 @@ package cluster
 import (
 	"fmt"
 
-	"github.com/ksonnet/ksonnet/client"
 	"github.com/ksonnet/ksonnet/pkg/app"
+	"github.com/ksonnet/ksonnet/pkg/client"
 	"github.com/ksonnet/ksonnet/pkg/pipeline"
 	"github.com/ksonnet/ksonnet/utils"
 	"github.com/pkg/errors"
