+++ pkg/cluster/delete.go
@@ -19,8 +19,8 @@ import (
 	"fmt"
 	"sort"
 
-	"github.com/ksonnet/ksonnet/client"
 	"github.com/ksonnet/ksonnet/pkg/app"
+	"github.com/ksonnet/ksonnet/pkg/client"
 	"github.com/ksonnet/ksonnet/utils"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
